Made Purchases a module
Note: There appear to be no tests
